backend/requirements.txt
# Core Flask dependencies
flask==3.0.0
flask-cors==4.0.0
flask-bcrypt==1.0.1
pyjwt==2.8.0
python-dotenv==1.0.0

# Vector database
qdrant-client==1.7.0

# ML/AI dependencies
sentence-transformers==2.2.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
transformers==4.35.2
numpy==1.24.3

# Google AI
google-generativeai==0.3.2
google-auth==2.23.4

# Supporting libraries
scikit-learn==1.3.2
pillow==10.1.0
huggingface-hub==0.19.4
tokenizers==0.15.0

# MLflow dependencies (if needed)
opentelemetry-proto==1.25.0

# Utilities
requests==2.31.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
urllib3==1.26.16

tqdm==4.66.1
filelock==3.13.1
fsspec==2023.12.2
pyyaml==6.0.1
regex==2023.12.25
joblib==1.3.2
threadpoolctl==3.2.0
sympy==1.13.1
setuptools==69.0.3
MarkupSafe==2.1.3
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
protobuf==4.25.1